Revert "type migration"
This reverts commit 77885522836e9f7b9e6c895843562f18863a0e0d.

+++ src/main/java/com/kel5/ecommerce/entity/Order.java
@@ -28,7 +28,7 @@ public class Order {
     private String address;
     private String whatsapp;
     private LocalDate orderDate;
-    private long totalAmount;
+    private float totalAmount;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
